backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pdfplumber==0.10.3
PyPDF2==3.0.1
sentence-transformers>=2.7.0
faiss-cpu==1.7.4
numpy<2.0
pydantic==2.5.3
python-dotenv==1.0.0
